requirements.txt
Flask==2.0.2
gunicorn==20.1.0
google-cloud-storage==1.43.0
Werkzeug==2.0.3
google-auth==2.0.2
google-api-python-client==2.30.0